pyproject.toml
[project]
name = "zamanbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "llama-index-llms-openai>=0.6.5",
    "llama-index>=0.14.5",
    "python-dotenv>=1.1.1",
]
[tool.mypy]
ignore_missing_imports = true
